+++ scifit/solvers/illdefined.py
@@ -1,3 +1,10 @@
 import numpy as np
 
 from scifit.interfaces.generic import FitSolverInterface
+
+
+class LinearSquaredSlopeSolver(FitSolverInterface):
+    @staticmethod
+    def model(x, a, b):
+        return np.power(a, 2) * x[:, 0] + b
+